+++ tv_shows_app/src/components/features/shows/ShowReviewSection/layouts/ShowReviewSection.desktop.tsx
@@ -1,14 +1,14 @@
 import { Flex, Text } from '@chakra-ui/react';
-import { IReview } from '../reviews/ReviewItem/ReviewItem';
-import { ReviewForm } from '../reviews/ReviewForm/ReviewForm';
-import { ReviewList } from '../reviews/ReviewList/ReviewList';
+import { IReview } from '../../../reviews/ReviewItem/ReviewItem';
+import { ReviewForm } from '../../../reviews/ReviewForm/ReviewForm';
+import { ReviewList } from '../../../reviews/ReviewList/ReviewList';
 
 interface IShowReviewSectionProps {
   reviews: Array<IReview>;
   showId: number;
 }
 
-export const ShowReviewSection = ({
+export const ShowReviewSectionDesktop = ({
   reviews,
   showId,
 }: IShowReviewSectionProps) => {
