+++ tv_shows_app/src/components/features/shows/ShowSection.tsx
@@ -1,14 +1,14 @@
 import { Flex, Image, Text, Box } from '@chakra-ui/react';
-import { ShowDetais } from './ShowDetails';
 import { IReview } from '../reviews/ReviewItem/ReviewItem';
 import { useEffect, useState } from 'react';
-import { ShowReviewSection } from './ShowReviewSection';
 import { IShow } from '@/typings/show';
 import { swrKeys } from '@/fetchers/swrKeys';
 import useSWR, { mutate } from 'swr';
 import { create } from 'domain';
 import { fetcher } from '@/fetchers/fetcher';
 import { LoadingScreen } from '@/components/shared/LoadingScreen/LoadingScreen';
+import { ShowReviewSection } from './ShowReviewSection/ShowReviewSection';
+import { ShowDetails } from './ShowDetails/ShowDetails';
 
 interface IShowSection {
   show: IShow;
@@ -28,10 +28,10 @@ export const ShowSection = ({ show }: IShowSection) => {
       backgroundColor="darkPurple"
       height="100%"
       padding="31px"
-      overflowY="auto"
+      overflow="auto"
     >
       <Flex flexDirection="column" alignItems="center" gap="61px">
-        <ShowDetais show={show} />
+        <ShowDetails show={show} />
         <ShowReviewSection reviews={data.reviews} showId={show.id} />
       </Flex>
     </Box>
