+++ tv_shows_app/src/components/shared/SidebarNavigation/layouts/SidebarNavigation.mobile.tsx
@@ -45,6 +45,7 @@ export const SidebarNavigationMobile = () => {
         <IconButton
           ml="auto"
           backgroundColor="transparent"
+          color="white"
           _hover={{ backgroundColor: 'transparent' }}
           colorScheme="teal"
           onClick={onOpen}
